clean stuff up
