+++ joplin/api/schema.py
@@ -14,7 +14,8 @@ from django_filters import FilterSet, OrderingFilter
 from wagtail.core.blocks import *
 from wagtail.documents.models import Document
 from wagtail.core.rich_text import expand_db_html
-from base.models import TranslatedImage, Theme, Contact, Location, PhoneNumber, ContactDayAndDuration
+from snippets.contact.models import Contact, PhoneNumber, ContactDayAndDuration
+from base.models import TranslatedImage, Theme
 
 from pages.topic_collection_page.models import TopicCollectionPage, JanisBasePageWithTopicCollections
 from pages.topic_page.models import TopicPage, TopicPageTopPage, JanisBasePageWithTopics
@@ -233,14 +234,6 @@ class TopicNode(DjangoObjectType):
     class Meta:
         model = TopicPage
         filter_fields = ['id', 'slug', 'live']
-        interfaces = [graphene.Node]
-
-
-class LocationNode(DjangoObjectType):
-    class Meta:
-        model = Location
-        filter_fields = ['id']
-        fields = '__all__'
         interfaces = [graphene.Node]
