+++ joplin/base/signals/aws_publish.py
@@ -8,8 +8,6 @@ from heroku3.models.build import Build
 
 import boto3
 
-from base.models import Contact, Location, Map
-
 import logging
 logger = logging.getLogger(__name__)
