+++ joplin/base/signals/janis_build_triggers.py
@@ -6,7 +6,8 @@ from graphene import Node
 from wagtail.core.models import Page
 from wagtail.admin.models import get_object_usage
 
-from base.models import Contact, Location, Map
+from base.models import Map
+from snippets.contact.models import Contact
 from pages.guide_page.models import GuidePage
 from groups.models import Department
 from wagtail.documents.models import Document
@@ -101,7 +102,6 @@ def find_pages_in_guides(changed_id):
 # we might want to log but not trigger a build? need some sort of queue
 @receiver(post_save, sender=Document)
 @receiver(post_save, sender=Contact)
-@receiver(post_save, sender=Location)
 @receiver(post_save, sender=Map)
 @receiver(post_save, sender=Department)
 def handle_post_save_signal(sender, **kwargs):
@@ -132,7 +132,6 @@ def page_unpublished_signal(sender, **kwargs):
 
 @receiver(post_delete, sender=Document)
 @receiver(post_delete, sender=Contact)
-@receiver(post_delete, sender=Location)
 @receiver(post_delete, sender=Map)
 def handle_post_delete_signal(sender, **kwargs):
     pages_global_ids = []
