+++ joplin/base/translation.py
@@ -2,7 +2,7 @@ from wagtail_modeltranslation.translation import register, TranslationOptions
 from wagtail.core.models import Page
 from wagtail.images.models import Image
 
-from base.models import TranslatedImage, Theme, Contact, Location, PhoneNumber, ContactDayAndDuration, Map
+from base.models import TranslatedImage, Theme, Map
 
 from pages.base_page.models import JanisBasePage
 from pages.topic_collection_page.models import TopicCollectionPage
