+++ joplin/settings.py
@@ -101,6 +101,7 @@ INSTALLED_APPS = [
     'pages.topic_collection_page',
     'pages.topic_page',
     'pages.home_page',
+    'snippets.contact',
 
 ]
