+++ joplin/snippets/contact/models.py
@@ -12,10 +12,8 @@ from wagtail.core.models import Orderable
 from phonenumber_field.modelfields import PhoneNumberField
 from phonenumber_field.widgets import PhoneNumberInternationalFallbackWidget
 
-from .day_and_duration import DayAndDuration
 
-from .constants import DEFAULT_MAX_LENGTH
-
+from base.models.constants import DEFAULT_MAX_LENGTH
 
 @register_snippet
 class Contact(ClusterableModel):
@@ -47,6 +45,49 @@ class Contact(ClusterableModel):
 
     def __str__(self):
         return self.name
+
+
+class DayAndDuration(ClusterableModel):
+    """
+    creates a model to choose day of week and hourly ranges
+    you can use this to define operating hours for a service or location
+    """
+    MONDAY = 'Monday'
+    TUESDAY = 'Tuesday'
+    WEDNESDAY = 'Wednesday'
+    THURSDAY = 'Thursday'
+    FRIDAY = 'Friday'
+    SATURDAY = 'Saturday'
+    SUNDAY = 'Sunday'
+    DAY_OF_WEEK_CHOICES = (
+        (MONDAY, 'Monday'),
+        (TUESDAY, 'Tuesday'),
+        (WEDNESDAY, 'Wednesday'),
+        (THURSDAY, 'Thursday'),
+        (FRIDAY, 'Friday'),
+        (SATURDAY, 'Saturday'),
+        (SUNDAY, 'Sunday'),
+    )
+
+    day_of_week = models.CharField(max_length=20, choices=DAY_OF_WEEK_CHOICES)
+    start_time = models.TimeField()
+    end_time = models.TimeField()
+
+    panels = [
+        FieldRowPanel(
+            children=[
+                FieldPanel('day_of_week', classname="col5"),
+                FieldPanel('start_time', classname="col3"),
+                FieldPanel('end_time', classname="col3"),
+
+            ],
+            classname="full"
+        ),
+    ]
+
+    def __str__(self):
+        return f'{self.day_of_week} {self.start_time} - {self.end_time}'
+
 
 
 class PhoneNumber(Orderable):
